+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
+- rpc: add smart cache ([#34](https://github.com/lamports-dev/alpamayo/pull/34))
+
 ### Breaking
 
 ## [0.4.0] - 2025-04-22